# Optimize snakecase processing