--- a/snake.go
+++ b/snake.go
@@ -1,26 +1,31 @@
 // Package snakecase - Super-Fast snake-case implementation.
 package snakecase
 
-const underscorByte = '_'
+const underscoreByte = '_'
 
 // Snakecase the given string.
 func Snakecase(s string) string {
 	idx := 0
-
-	for ; idx < len(s); idx++ {
-		if isAlphanumeric(s[idx]) {
-			break
-		}
-	}
-	s = s[idx:]
-	idx = 0
+	hasLower := false
+	hasUnderscore := false
+	lowercaseSinceUnderscore := false
 
 	// loop through all good characters:
 	// - lowercase
 	// - digit
 	// - underscore (as long as the next character is lowercase or digit)
 	for ; idx < len(s); idx++ {
-		if (isLower(s[idx]) || isDigit(s[idx])) || (s[idx] == underscorByte && idx < len(s)-1 && (isLower(s[idx+1]) || isDigit(s[idx+1]))) {
+		if isLower(s[idx]) {
+			hasLower = true
+			if hasUnderscore {
+				lowercaseSinceUnderscore = true
+			}
+			continue
+		} else if isDigit(s[idx]) {
+			continue
+		} else if s[idx] == underscoreByte && idx > 0 && idx < len(s)-1 && (isLower(s[idx+1]) || isDigit(s[idx+1])) {
+			hasUnderscore = true
+			lowercaseSinceUnderscore = false
 			continue
 		}
 		break
@@ -29,15 +34,22 @@ func Snakecase(s string) string {
 	if idx == len(s) {
 		return s // no changes needed, can just borrow the string
 	}
-	//panic(s[0:idx])
+
 	// if we get here then we must need to manipulate the string
 	b := make([]byte, 0, 64)
-
-	// handles digit followed by an uppercase character, but only if the digit is not the leading character
-	if (idx == 1 && isDigit(s[idx-1])) || (idx > 1 && isDigit(s[idx-1]) && !isAlphanumeric(s[idx-2])) {
-		idx--
-	}
 	b = append(b, s[:idx]...)
+
+	if isUpper(s[idx]) && (!hasLower || hasUnderscore && !lowercaseSinceUnderscore) {
+		for idx < len(s) && (isUpper(s[idx]) || isDigit(s[idx])) {
+			b = append(b, asciiLowercaseArray[s[idx]])
+			idx++
+		}
+
+		for idx < len(s) && (isLower(s[idx]) || isDigit(s[idx])) {
+			b = append(b, s[idx])
+			idx++
+		}
+	}
 
 	for idx < len(s) {
 		if !isAlphanumeric(s[idx]) {
@@ -46,7 +58,7 @@ func Snakecase(s string) string {
 		}
 
 		if len(b) > 0 {
-			b = append(b, underscorByte)
+			b = append(b, underscoreByte)
 		}
 
 		for idx < len(s) && (isUpper(s[idx]) || isDigit(s[idx])) {