**Details**
I was messing around with [Rust](https://www.rust-lang.org/) which has the ability to borrow string data and makes you think a little bit differently and this is a new algorithm I came up with as a result; it avoids allocating and manipulating the string unless necessary.

Probably would want to test in stage first with a larger set of data, but it's here when ready 😄 

**Comparison**
```
benchcmp old.txt new.txt
benchmark                                     old ns/op     new ns/op     delta
BenchmarkSnakecaseBorrowedLong-8              84.9          19.0          -77.62%
BenchmarkSnakecaseBorrowedSimple-8            40.0          8.95          -77.62%
BenchmarkSnakecaseOwnedUnicode-8              41.4          36.7          -11.35%
BenchmarkSnakecaseOwnedLong-8                 71.8          57.2          -20.33%
BenchmarkSnakecaseOwnedLongSpecialChars-8     102           77.1          -24.41%
BenchmarkSnakecaseOwnedSimple-8               39.2          34.1          -13.01%
BenchmarkSnakecaseOwnedUnicode2-8             102           76.5          -25.00%

benchmark                                     old MB/s     new MB/s     speedup
BenchmarkSnakecaseBorrowedLong-8              388.67       1735.43      4.47x
BenchmarkSnakecaseBorrowedSimple-8            274.70       1229.60      4.48x
BenchmarkSnakecaseOwnedUnicode-8              217.49       245.46       1.13x
BenchmarkSnakecaseOwnedLong-8                 404.05       507.17       1.26x
BenchmarkSnakecaseOwnedLongSpecialChars-8     292.22       388.86       1.33x
BenchmarkSnakecaseOwnedSimple-8               280.62       322.92       1.15x
BenchmarkSnakecaseOwnedUnicode2-8             252.78       339.81       1.34x

benchmark                                     old allocs     new allocs     delta
BenchmarkSnakecaseBorrowedLong-8              1              0              -100.00%
BenchmarkSnakecaseBorrowedSimple-8            1              0              -100.00%
BenchmarkSnakecaseOwnedUnicode-8              1              1              +0.00%
BenchmarkSnakecaseOwnedLong-8                 1              1              +0.00%
BenchmarkSnakecaseOwnedLongSpecialChars-8     1              1              +0.00%
BenchmarkSnakecaseOwnedSimple-8               1              1              +0.00%
BenchmarkSnakecaseOwnedUnicode2-8             1              1              +0.00%

benchmark                                     old bytes     new bytes     delta
BenchmarkSnakecaseBorrowedLong-8              48            0             -100.00%
BenchmarkSnakecaseBorrowedSimple-8            16            0             -100.00%
BenchmarkSnakecaseOwnedUnicode-8              3             3             +0.00%
BenchmarkSnakecaseOwnedLong-8                 48            48            +0.00%
BenchmarkSnakecaseOwnedLongSpecialChars-8     32            32            +0.00%
BenchmarkSnakecaseOwnedSimple-8               16            16            +0.00%
BenchmarkSnakecaseOwnedUnicode2-8             3             3             +0.00%
```

**Old**
```
BenchmarkSnakecaseBorrowedLong-8            	20000000	        84.9 ns/op	 388.67 MB/s	      48 B/op	       1 allocs/op
BenchmarkSnakecaseBorrowedSimple-8          	30000000	        40.0 ns/op	 274.70 MB/s	      16 B/op	       1 allocs/op
BenchmarkSnakecaseOwnedUnicode-8            	30000000	        41.4 ns/op	 217.49 MB/s	       3 B/op	       1 allocs/op
BenchmarkSnakecaseOwnedLong-8               	20000000	        71.8 ns/op	 404.05 MB/s	      48 B/op	       1 allocs/op
BenchmarkSnakecaseOwnedLongSpecialChars-8   	20000000	       102 ns/op	 292.22 MB/s	      32 B/op	       1 allocs/op
BenchmarkSnakecaseOwnedSimple-8             	50000000	        39.2 ns/op	 280.62 MB/s	      16 B/op	       1 allocs/op
BenchmarkSnakecaseOwnedUnicode2-8           	20000000	       102 ns/op	 252.78 MB/s	       3 B/op	       1 allocs/op
```

**New**
```
BenchmarkSnakecaseBorrowedLong-8            	100000000	        19.0 ns/op	1735.43 MB/s	       0 B/op	       0 allocs/op
BenchmarkSnakecaseBorrowedSimple-8          	200000000	         8.95 ns/op	1229.60 MB/s	       0 B/op	       0 allocs/op
BenchmarkSnakecaseOwnedUnicode-8            	50000000	        36.7 ns/op	 245.46 MB/s	       3 B/op	       1 allocs/op
BenchmarkSnakecaseOwnedLong-8               	30000000	        57.2 ns/op	 507.17 MB/s	      48 B/op	       1 allocs/op
BenchmarkSnakecaseOwnedLongSpecialChars-8   	20000000	        77.1 ns/op	 388.86 MB/s	      32 B/op	       1 allocs/op
BenchmarkSnakecaseOwnedSimple-8             	50000000	        34.1 ns/op	 322.92 MB/s	      16 B/op	       1 allocs/op
BenchmarkSnakecaseOwnedUnicode2-8           	20000000	        76.5 ns/op	 339.81 MB/s	       3 B/op	       1 allocs/op
```