@joeybloggs 2019-01-22
Thanks @achille-roussel for the feedback

I saw that number behaviour change and made a fix.

Oh I totally agree under no circumstances should this be merged without extremely thourough testing :)

will make a few adjustments based on your suggestions:)