@deankarn 2019-05-20
@dominicbarnes @achille-roussel I was finally able to achieve feature parity which yields the following:
```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkUnchangedLong-8                    95.2          37.6          -60.50%
BenchmarkUnchangedSimple-8                  46.1          14.5          -68.55%
BenchmarkModifiedUnicode-8                  36.4          41.4          +13.74%
BenchmarkModifiedLong-8                     85.6          81.9          -4.32%
BenchmarkModifiedLongSpecialChars-8         128           127           -0.78%
BenchmarkModifiedSimple-8                   42.3          45.0          +6.38%
BenchmarkModifiedUnicode2-8                 71.7          79.4          +10.74%
BenchmarkLeadingUnderscoresDigitUpper-8     37.8          39.4          +4.23%
BenchmarkDigitUpper-8                       33.4          28.9          -13.47%
BenchmarkDigitUpper2-8                      83.5          84.8          +1.56%

benchmark                                   old MB/s     new MB/s     speedup
BenchmarkUnchangedLong-8                    346.67       878.58       2.53x
BenchmarkUnchangedSimple-8                  238.83       757.91       3.17x
BenchmarkModifiedUnicode-8                  247.09       217.28       0.88x
BenchmarkModifiedLong-8                     338.64       354.06       1.05x
BenchmarkModifiedLongSpecialChars-8         233.89       234.43       1.00x
BenchmarkModifiedSimple-8                   259.76       244.48       0.94x
BenchmarkModifiedUnicode2-8                 362.85       327.45       0.90x
BenchmarkLeadingUnderscoresDigitUpper-8     158.70       152.21       0.96x
BenchmarkDigitUpper-8                       149.82       172.80       1.15x
BenchmarkDigitUpper2-8                      203.71       200.45       0.98x

benchmark                                   old allocs     new allocs     delta
BenchmarkUnchangedLong-8                    1              0              -100.00%
BenchmarkUnchangedSimple-8                  1              0              -100.00%
BenchmarkModifiedUnicode-8                  1              1              +0.00%
BenchmarkModifiedLong-8                     1              1              +0.00%
BenchmarkModifiedLongSpecialChars-8         1              1              +0.00%
BenchmarkModifiedSimple-8                   1              1              +0.00%
BenchmarkModifiedUnicode2-8                 1              1              +0.00%
BenchmarkLeadingUnderscoresDigitUpper-8     1              1              +0.00%
BenchmarkDigitUpper-8                       1              1              +0.00%
BenchmarkDigitUpper2-8                      1              1              +0.00%

benchmark                                   old bytes     new bytes     delta
BenchmarkUnchangedLong-8                    48            0             -100.00%
BenchmarkUnchangedSimple-8                  16            0             -100.00%
BenchmarkModifiedUnicode-8                  3             3             +0.00%
BenchmarkModifiedLong-8                     48            48            +0.00%
BenchmarkModifiedLongSpecialChars-8         32            32            +0.00%
BenchmarkModifiedSimple-8                   16            16            +0.00%
BenchmarkModifiedUnicode2-8                 3             3             +0.00%
BenchmarkLeadingUnderscoresDigitUpper-8     5             5             +0.00%
BenchmarkDigitUpper-8                       5             5             +0.00%
BenchmarkDigitUpper2-8                      32            32            +0.00%
```

### Testing
I tested by randomly generating 1 billion random strings, running them through the old logic and then verifying the new logic generated the exact same output; here are the programs for reference:
Generate
```go
package main

import (
	"fmt"
	"math/rand"
	"time"

	snakecase "github.com/segmentio/go-snakecase"
)

const stringLen = 35

var characters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890_@ẞ•¶§ƒ∆˚¬")

func init() {
	rand.Seed(time.Now().UnixNano())
}

func getRandomString(b []rune) []rune {
	num := rand.Intn(stringLen)
	if num == 0 {
		num++
	}
	for i := 0; i < num; i++ {
		b = append(b, characters[rand.Intn(len(characters))])
	}
	return b
}

func main() {
	var b []rune
	for i := 0; i < 1000000000; i++ {
		b = getRandomString(b[:0])
		s := string(b)
		fmt.Printf("%s:%s\n", s, snakecase.Snakecase(s))
	}
}
```
Verify
```go
package main

import (
	"bufio"
	"fmt"
	"log"
	"os"
	"strings"

	snakecase "github.com/segmentio/go-snakecase"
)

func main() {
	f, err := os.Open("test.txt")
	if err != nil {
		log.Fatal(err)
	}
	defer f.Close()
	r := bufio.NewReader(f)
	s := bufio.NewScanner(r)
	var sc string
	for s.Scan() {
		st := strings.Split(s.Text(), ":")
		sc = snakecase.Snakecase(st[0])
		if sc != st[1] {
			fmt.Printf("%s:%s:%s\n", st[0], st[1], sc)
		}
	}
}
```