@achille-roussel 2019-05-23
It looks like some benchmarks are running slower with this change, would it be worth testing if this is actually yielding positive returns on a real data set?