@deankarn 2019-05-23
@achille-roussel on the 1 billion string test which consists of many unicode and changed strings and it still yielded a small increase is performance and that test is almost the worst cases real-world example.

99% of the time from my side it will hit the new optimized paths.