@deankarn 2019-05-23
@achille-roussel thanks 👍 

agree with having more than one approver and opinion!

Next step: https://www.rust-lang.org/ ;)
(although my rust version does not seem to be quite as fast, Go does a great job with string manipulation)